move events back to textfile service
