+++ src/vs/workbench/services/textfile/common/textfiles.ts
@@ -4,10 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { URI } from 'vs/base/common/uri';
-import { Event } from 'vs/base/common/event';
+import { Event, IWaitUntil } from 'vs/base/common/event';
 import { IDisposable } from 'vs/base/common/lifecycle';
 import { IEncodingSupport, ConfirmResult, IRevertOptions, IModeSupport } from 'vs/workbench/common/editor';
-import { IBaseStatWithMetadata, IFileStatWithMetadata, IReadFileOptions, IWriteFileOptions, FileOperationError, FileOperationResult } from 'vs/platform/files/common/files';
+import { IBaseStatWithMetadata, IFileStatWithMetadata, IReadFileOptions, IWriteFileOptions, FileOperationError, FileOperationResult, FileOperation } from 'vs/platform/files/common/files';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
 import { ITextEditorModel } from 'vs/editor/common/services/resolverService';
 import { ITextBufferFactory, ITextModel, ITextSnapshot } from 'vs/editor/common/model';
@@ -27,6 +27,16 @@ export interface ITextFileService extends IDisposable {
 	readonly onFilesAssociationChange: Event<void>;
 
 	readonly isHotExitEnabled: boolean;
+
+	/**
+	 * An event that is fired before attempting a certain file operation.
+	 */
+	readonly onWillRunOperation: Event<FileOperationWillRunEvent>;
+
+	/**
+	 * An event that is fired after a file operation has been performed.
+	 */
+	readonly onDidRunOperation: Event<FileOperationDidRunEvent>;
 
 	/**
 	 * Access to the manager of text file editor models providing further methods to work with them.
@@ -144,6 +154,35 @@ export interface ITextFileService extends IDisposable {
 	 */
 	getAutoSaveConfiguration(): IAutoSaveConfiguration;
 }
+
+
+export class FileOperationWillRunEvent implements IWaitUntil {
+
+	constructor(
+		private _thenables: Promise<any>[],
+		readonly operation: FileOperation,
+		readonly target: URI,
+		readonly source?: URI | undefined
+	) { }
+
+	waitUntil(thenable: Promise<any>): void {
+		if (Object.isFrozen(this._thenables)) {
+			throw new Error('waitUntil cannot be used aync');
+		}
+		this._thenables.push(thenable);
+	}
+}
+
+
+export class FileOperationDidRunEvent {
+
+	constructor(
+		readonly operation: FileOperation,
+		readonly target: URI,
+		readonly source?: URI | undefined
+	) { }
+}
+
 
 export interface IReadTextFileOptions extends IReadFileOptions {
