+++ src/vs/workbench/test/workbenchTestServices.ts
@@ -923,8 +923,6 @@ export class TestFileService implements IFileService {
 
 	public _serviceBrand: undefined;
 
-	readonly onWillRunOperation = Event.None;
-
 	private readonly _onFileChanges: Emitter<FileChangesEvent>;
 	private readonly _onAfterOperation: Emitter<FileOperationEvent>;
